requirements.txt
aiohttp
aiofiles
beautifulsoup4
python-dotenv
PyGithub
requests
tqdm
rich
textual